calculate ttl in other rust package
